# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend developer test",
        "version": "1.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "UserController"
                ],
                "summary": "Users list",
                "operationId": "4cfb5d1cbc90fb1fc4dfb569aef1b6e9",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": [
                                            {
                                                "_id": "620e5cbe2709000014007762",
                                                "email": "eve.holt@reqres.in",
                                                "updated_at": "2022-02-17T23:10:36.439000Z",
                                                "created_at": "2022-02-17T14:33:33.991000Z"
                                            },
                                            {
                                                "_id": "620e9ff12709000014007765",
                                                "email": "marif@gmail.com",
                                                "updated_at": "2022-02-17T22:51:48.836000Z",
                                                "created_at": "2022-02-17T19:20:17.786000Z"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Unauthorized."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "UserController"
                ],
                "summary": "Users detail",
                "operationId": "2b4dac75507ff9ba001d01e6cd813288",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "_id": "620e5cbe2709000014007762",
                                            "email": "eve.holt@reqres.in",
                                            "updated_at": "2022-02-17T23:10:36.439000Z",
                                            "created_at": "2022-02-17T14:33:33.991000Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "User tidak ditemukan"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Unauthorized."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/create": {
            "post": {
                "tags": [
                    "UserController"
                ],
                "summary": "User add",
                "operationId": "6f3d7846f3ba79f1d71d00bec3cd6c6a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "john@mail.com",
                                    "password": "1234567"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "email": "john@mail.com",
                                            "updated_at": "2022-02-17T22:39:19.581000Z",
                                            "created_at": "2022-02-17T22:39:19.581000Z",
                                            "_id": "620ece972709000014007766"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "errors": {
                                                "email": [
                                                    "The email field is required."
                                                ],
                                                "password": [
                                                    "The password field is required."
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Unauthorized."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}/update": {
            "put": {
                "tags": [
                    "UserController"
                ],
                "summary": "User update",
                "operationId": "3ab709efe471ba01cbb8feff94fd03f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "john@mail.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "_id": "620e9ff12709000014007765",
                                            "email": "john@mail.com",
                                            "updated_at": "2022-02-17T22:51:48.836000Z",
                                            "created_at": "2022-02-17T19:20:17.786000Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "User tidak ditemukan"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "errors": {
                                                "email": [
                                                    "The email field is required."
                                                ],
                                                "password": [
                                                    "The password field is required."
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Unauthorized."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}/delete": {
            "delete": {
                "tags": [
                    "UserController"
                ],
                "summary": "User delete",
                "operationId": "df79c8aef908bd01da604168698ea172",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "success"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "User tidak ditemukan"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Unauthorized."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}